services/FeedService/pyproject.toml
[project]
name = "feedservice"
version = "0.1.0"
description = "Feed and algorithmic service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "redis>=5.0.0",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.135.2",
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy[asyncio]>=2.0.48",
    "uvicorn>=0.42.0",
    "pyjwt[crypto]>=2.12.1",
    "ruff>=0.15.8",
]